django/requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
billiard==3.3.0.23
celery==3.1.25
coreapi==2.3.0
coreschema==0.0.4
Django==1.11
django-celery==3.2.1
djangorestframework==3.6.2
itypes==1.1.0
Jinja2==2.9.6
kombu==3.0.37
MarkupSafe==1.0
olefile==0.44
openapi-codec==1.3.1
packaging==16.8
Pillow==4.1.0
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
requests==2.13.0
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
vine==1.1.3
Wand==0.4.4
